Move NewDefaultCell type to src/types directory
