+++ ui/src/dashboards/utils/cellGetters.ts
@@ -1,6 +1,6 @@
 import {NEW_DEFAULT_DASHBOARD_CELL} from 'src/dashboards/constants'
-import {Cell, CellType, Dashboard} from 'src/types/dashboards'
-import {NewDefaultCell, UNTITLED_GRAPH} from 'src/dashboards/constants'
+import {Cell, CellType, Dashboard, NewDefaultCell} from 'src/types/dashboards'
+import {UNTITLED_GRAPH} from 'src/dashboards/constants'
 
 const getMostCommonValue = (values: number[]): number => {
   const results = values.reduce(
